backend/requirements.txt
asgiref==3.8.1
certifi==2024.2.2
cffi==1.16.0
channels==4.2.2
channels-redis==4.2.1
charset-normalizer==3.3.2
coreapi==2.3.3
coreschema==0.0.4
cryptography==42.0.7
defusedxml==0.8.0rc2
Django==5.0.6
django-cors-headers==4.6.0
django-templated-mail==1.1.1
djangorestframework==3.15.1
djangorestframework-simplejwt==5.3.1
djoser==2.2.2
idna==3.7
itypes==1.2.0
Jinja2==3.1.6
MarkupSafe==3.0.2
msgpack==1.0.8
oauthlib==3.2.2
Pillow==11.2.1
pycparser==2.22
PyJWT==2.8.0
python3-openid==3.2.0
pytz==2024.2
PyYAML==6.0.2
redis==6.1.0
requests==2.32.2
requests-oauthlib==2.0.0
social-auth-app-django==5.4.1
social-auth-core==4.5.4
sqlparse==0.5.0
tzdata==2024.2
uritemplate==4.1.1
urllib3==2.2.1
